Lưu trạng thái cửa hàng
