+++ app/src/main/res/layout/activity_gio_hang.xml
@@ -11,7 +11,7 @@
 
         <ImageButton
             android:id="@+id/imgButtonLeft"
-            android:layout_width="46dp"
+            android:layout_width="50dp"
             android:layout_height="50dp"
             android:background="@color/appBackground"
             android:gravity="left"
@@ -46,15 +46,7 @@
         android:layout_height="0dp"
         android:layout_weight="1"
         android:gravity="center"
-        android:orientation="vertical">
-
-        <Button
-            android:id="@+id/btnContinue"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@color/appBackground"
-            android:text="Tiếp tục mua hàng"></Button>
-    </LinearLayout>
+        android:orientation="vertical"/>
 
     <LinearLayout
         android:layout_width="match_parent"
